split test cases to run more tests in parallel
